Add dragEvent file - v0.6.15
