+++ src/index.html
@@ -9,7 +9,7 @@
 
 <body>
   <div class="container">
-    <div id="slider2" class="slider">
+    <div id="slider" class="slider">
       <div class="wrapper">
         <div class="slides">
           <span class="slide">Slide 1</span>
