+++ src/scripts/index.js
@@ -1,14 +1,13 @@
 import '../styles/index.scss';
-// import config from './config';
 import slide from './events';
 
 const slideConfig = {
-	slider: "#slider2",//add user
+	slider: "#slider",//add user
 	threshold: 50, // add user
 	infinite: false, //add user
 	responsive: { //add user
 		0: {
-			items: 2
+			items: 1.5
 		},
 		560: {
 			items: 2.5
