+++ src/scripts/utils.js
@@ -1,71 +1,3 @@
-import config from "./config";
-
-export const caroueslTouchStart = e => e.touches[0].clientX;
-
-export const caroueslDragAction = params => {
-  const { e, dragEnd, dragAction } = params;
-  document.onmouseup = dragEnd;
-  document.onmousemove = dragAction;
-  return e.clientX;
-};
-
-export const dragActionTouchmovePosX2 = params => {
-  const { e, posX1 } = params;
-  return posX1 - e.touches[0].clientX;
-};
-
-export const dragActionTouchmovePosX1 = e => e.touches[0].clientX;
-
-export const dragActionMousemove = params => {
-  const { posX1, e } = params;
-  return posX1 - e.clientX;
-};
-
-export const dragActionMousemovePosX1 = e => e.clientX;
-
-export const dragActionCalcPosition = params => {
-  const {
-    sliderItems,
-    posX2,
-    slidesLength,
-    sliderItemWidth,
-    countItem,
-    slideSize,
-    sliderMainWidth,
-    infinite,
-    threshold
-  } = params;
-
-  // stop scroll when firstItem go to lastItem on drag
-  if (getTranslate3d(sliderItems) - posX2 > -sliderItemWidth * countItem + threshold) {
-    return false;
-  }
-
-  // stop scroll when lastItem go to firstItem on drag
-  if (
-    getTranslate3d(sliderItems) - posX2 <=
-    calcFinalItemPosition({
-      slideSize,
-      slidesLength,
-      sliderMainWidth,
-      countItem,
-      infinite
-    }) -
-    threshold
-  ) {
-    return false;
-  }
-
-  sliderItems.style["transform"] = setTranslate3d(
-    getTranslate3d(sliderItems) - posX2
-  );
-};
-
-export const mouseEventNull = () => {
-  document.onmouseup = null;
-  document.onmousemove = null;
-};
-
 export const sliderItemsAddClass = sliderItems => {
   sliderItems.classList.add("shifting");
   return false;
@@ -343,7 +275,7 @@ export const calcSliderChildWidth = (params) => {
   }
   const mainWidthTruncItem = sliderClientWidth(slider) / itemsTrunc;
   let decriseWithForEachItems = mainWidthTruncItem / itemsTrunc / itemsTrunc;
-  if (items > 1 && items < 2) {
+  if (responsiveItemCount > 1 && responsiveItemCount < 2) {
     decriseWithForEachItems = (sliderClientWidth(slider) / itemsTrunc) * 0.25;
   }
